Refactor Arduino code structure moving evreything on esp32: remove old communication files, add new sensor and positioning logic, and implement setup functions for sensor configuration.
