--- src/arduino/comunication.h
@@ -1,9 +1,0 @@
-#ifndef COMUNICATION_H
-#define COMUNICATION_H
-
-#include <Arduino.h>
-
-void I2C_DataSend();
-void maxReadGet(int setupInfo);
-
-#endif
