+++ src/centering_control.cpp
