+++ src/config.h
@@ -2,29 +2,26 @@
 #define CONFIG_H
 
 #include <Arduino.h>
+#include <Preferences.h>
 
-#define MYADDRESS 0x10
 #define ND -9999
 #define DATATIMES 21
-#define SIMULATION 1 //set this to 0 to get data from ESP32, 1 to simulate data for MAXdistance = 80cm
-
 
 // Pin definitions
-#if defined(ARDUINO_AVR_UNO)
+#if defined(ESP32_DEV_BOARD)
     const int trigPins[3] = {2, 4, 6};
     const int echoPins[3] = {3, 5, 7};
-#elif defined(ARDUINO_AVR_MEGA2560)
-    const int trigPins[3] = {22, 24, 26};
-    const int echoPins[3] = {23, 25, 27};
 #else
     #error "Unsupported board"
 #endif
 
-// Global variables (shared)
+//global variables (shared)
+extern Preferences storage;
+extern float Ax, Ay, Bx, By, Cx, Cy;
+extern float Dx, Dy;
 extern int MAXdistance;
 extern int MINdistance;
-extern int toSendDistances[3];
-extern volatile bool maxKnown;
+extern int Distances[3];
 const int MAXtimeout = 15000; // Maximum time to wait for echo (in microseconds)
 
 #endif
