+++ src/drone.cpp
